hitung_gaji/requirements.txt
absl-py==2.3.1
antiorm==1.2.1
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
DateTime==5.5
db==0.1.1
Django==5.2.3
et_xmlfile==2.0.0
exc==0.92
fanstatic==1.6
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fonttools==4.58.1
greenlet==3.2.2
gTTS==2.5.4
idna==3.10
include==0.2.2
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
js==1.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.3
mysql-connector-python==9.6.0
mysqlclient==2.2.7
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==4.25.8
pycparser==2.23
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scipy==1.16.2
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.2
SQLAlchemy==2.0.41
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
WebOb==1.8.9
Werkzeug==3.1.3
zope.interface==7.2
